requirements.txt
Django==1.10.6         	# Framework
mysqlclient==1.3.10    	# Python interface to MySQL
pytz==2017.2           	# TimeZone used for Django < 1.11
Pillow==4.2.1          	# Python Image Library (PIL)
xlrd==1.1.0            	# Reading data and formatting information from Excel files, whether they are .xls or .xlsx files
celery==4.1.0          	# Task queue
channels==1.1.8			# WebSockets and HTTP2 server for Django
asgi_rabbitmq==0.4.1	# RabbitMQ backend for ASGI
